Add test harness target for iOS apps.
Summary:
For a given library, the catalog app target depends on the lib source and the unit test target depends on the unit test source. CocoaPods generates one .framework for both: one has just the lib source, the other has the lib source + unit test source. If we attempt to build the unit test target in the Catalog app as a host, the Catalog app's .framework wins and our unit tests fail to load their source. This can be resolved by creating a TestHarness app which has no dependencies, but simply facilitates the running of the unit tests (what we want).

This diff updates the templates to reflect this new setup.

Reviewers: markwei, O2 Material Motion, O4 Material Motion Apple platform reviewers

Reviewed By: markwei, O2 Material Motion, O4 Material Motion Apple platform reviewers

Subscribers: markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D2064